+++ changelog.md
@@ -1,5 +1,12 @@
 # WinRM-Elevated Gem Changelog
 
+# 0.4.0
+- Initialize `Elevated::Runner` with a `CommandExecutor` instead of a `WinrmService` client
+- Run commands from newer winrm executor
+- Use latest winrm-fs 0.4.2
+- Allow task to run as a service account
+- Provide an artificially long timeout to the task to keep the task from dying after 60 seconds
+
 # 0.3.0
 - [Name Powershell Script and Log Files Uniquely](https://github.com/WinRb/winrm-elevated/pull/6)